feat(chat-demo): refactor chat demo section and introduce new components
- Refactored the ChatDemoSection to utilize new components: ChatWindow, ChatMessages, ChatHeader, and ChatInput, enhancing modularity and readability.
- Introduced CapabilitiesList component to display assistant capabilities dynamically.
- Implemented useChatAnimation hook for managing chat message animations and visibility.
- Removed hardcoded chat messages and mascot images, improving flexibility and maintainability.
