+++ src/components/pages/splash-page/splash-sections/chat-demo/index.ts
@@ -1,0 +1,8 @@
+export { ChatMessageBubble } from "./ChatMessageBubble";
+export type { ChatMessageBubbleProps } from "./ChatMessageBubble";
+export { ChatHeader } from "./ChatHeader";
+export { ChatInput } from "./ChatInput";
+export { ChatMessages } from "./ChatMessages";
+export { ChatWindow } from "./ChatWindow";
+export { CapabilitiesList } from "./CapabilitiesList";
+export { useChatAnimation } from "./useChatAnimation";
